Use ptrdiff_t

+++ apps/openmw/mwgui/windowbase.cpp
@@ -26,7 +26,7 @@ size_t MWGui::wrap(size_t index, size_t max, int delta)
             return 0;
         return index + absDelta;
     }
-    size_t absDelta = static_cast<size_t>(-static_cast<long>(delta));
+    size_t absDelta = static_cast<size_t>(-static_cast<ptrdiff_t>(delta));
     if (index >= absDelta)
         return index - absDelta;
     else if (max == 0)
